"Prodigy_ML_TASK 01
Above mentioned ipynb file is related to machine learning project.
This is prediction of insurance premium charges which is taken from Kaggle.  
 
 Description about project:-
          
Insurance premiums are often based on various factors that in the end decide the amount that will be covered from the insurance company. So this given set of historical data for an organization's customers and the respective charges that were levied upon the insurance company.

The objective of proposed work is to predict the insurance charges of a person and identify those patients with health insurance policy and medical details weather they have any health issues or not.

The data gives you the information about the users including their age, sex, bmi, hospitalization history, annual income, etc. Analyze and gather insights from the data and create a linear regression model that will best predict the insurance charges for a new set of data.

This kind of model is useful for insurance companies to determine the yearly insurance premium charges for a person.

In this project:
Firstly, pass all essential libraries and upload the data which is taken from Kaggle. After uploading the data, EDA (Exploratory data analysis) is the next step in which removing duplicates, null values, outliers, multicollinearity all steps done systematically. after EDA part the data is clean and filtered and ready for visualization with some informative graphs. After visualization the next part is Feature selection which is the LINEAR REGRESSION in which the data divided into two variables and preparing for accuracy which is 85.22 percent.